Fix some issues with the GTFS stop_times and shapes that caused strange behavior after the latest update from PATS in August 2015.
Despite the previous data passing validation using the various tools, OTP has it's own parser with the following quirks:

Since Bullrunner routes operate as large loops, in order to coerce OTP to accept the full loop shape correctly the starting stop is duplicated at the end of the route.

This works for some routes, but others have shapes (from Syncromatics) with a start/end segment in between stops.  When building a graph, OTP is unable to process a stop with a given shape segment and considers it invalid - defaulting instead to straight lines.

To fix it, I needed to carefully ensure the end of the shape for these routes ends exactly at the ending stop (now also the beginning stop).

Without making those changes, the route shape polylines would awkwardly end at the end of a route with no continuation to the beginning again.